+++ bips/src/test/java/io/github/novacrypto/mnemonics/NumericEntryModelCoverageTests.kt
@@ -1,0 +1,2 @@
+package io.github.novacrypto.mnemonics
+
